+++ changelog.md
@@ -1,5 +1,7 @@
 # ChangeLog
 
+* Fix gmail XOAUTH2 scope, thanks to @sherryl4george
+
 ## Version 5.2.14 (Nov 1st 2015)
 * Allow addresses with IDN (Internationalized Domain Name) in PHP 5.3+, thanks to @fbonzon
 * Allow access to POP3 errors